backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-environ>=0.10.0
django-extensions>=3.2.0

# Database
psycopg2-binary>=2.9.0
redis>=4.5.0

# Authentication & Security
djangorestframework-simplejwt>=5.2.0
django-allauth>=0.54.0
cryptography>=40.0.0

# Background Tasks
celery>=5.2.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# API Documentation
drf-yasg>=1.21.0

# Time Series & Analytics
pandas>=2.0.0
numpy>=1.24.0

# HTTP Client for AI integration
requests>=2.28.0
httpx>=0.24.0

# Filtering and utilities
django-filter>=23.2
python-dateutil>=2.8.0
Pillow>=9.5.0
django-redis>=5.2.0

# Development & Testing
pytest>=7.3.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
factory-boy>=3.2.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Production
gunicorn>=20.1.0
whitenoise>=6.4.0